requirements.txt
attrs==24.2.0
python-dotenv==1.0.1
screeninfo==0.8.1
selenium==4.26.1
urllib3==2.2.3
websocket-client==1.8.0
